BIG changes
- restructured models.py for BCO and Prefix
- modify many other files to accommodate above change
- added initial users to `bca_api/api/fixtures` as `api_initial.sql`
Changes to be committed:
	modified:   .gitignore
	modified:   bco_api/api/admin.py
	modified:   bco_api/api/apps.py
	new file:   bco_api/api/fixtures/api_initial.json
	new file:   bco_api/api/fixtures/api_initial.sql
	deleted:    bco_api/api/fixtures/bootstrap.json
	deleted:    bco_api/api/fixtures/metafixtures
	deleted:    bco_api/api/fixtures/metafixtures.json
	new file:   bco_api/api/migrations/0004_auto_20220224_1946.py
	modified:   bco_api/api/models.py
	new file:   bco_api/api/scripts/groups.py
	modified:   bco_api/api/scripts/method_specific/GET_activate_account.py
	modified:   bco_api/api/scripts/method_specific/POST_api_accounts_new.py
	modified:   bco_api/api/scripts/method_specific/POST_api_groups_delete.py
	modified:   bco_api/api/scripts/method_specific/POST_api_objects_drafts_create.py
	modified:   bco_api/api/scripts/method_specific/POST_api_objects_drafts_publish.py
	modified:   bco_api/api/scripts/method_specific/POST_api_objects_drafts_token.py
	modified:   bco_api/api/scripts/method_specific/POST_api_prefixes_create.py
	modified:   bco_api/api/scripts/method_specific/POST_api_prefixes_delete.py
	modified:   bco_api/api/scripts/method_specific/POST_api_prefixes_modify.py
	modified:   bco_api/api/scripts/method_specific/POST_api_prefixes_permissions_set.py
	modified:   bco_api/api/scripts/utilities/DbUtils.py
	modified:   bco_api/api/scripts/utilities/UserUtils.py
	modified:   bco_api/api/signals.py
	modified:   bco_api/api/views.py
	modified:   requirements.txt

+++ bco_api/api/models.py
@@ -15,21 +15,70 @@ from django.contrib.auth.models import Group, Permission, User
 from django.db.models.signals import post_save, post_delete, pre_delete
 from django.dispatch import receiver
 import django.db.utils as PermErrors
-
 # Referencing models.
 from django.contrib.contenttypes.models import ContentType
-
 # Issue with timezones.
 # Source: https://stackoverflow.com/a/32411560
 from django.utils import timezone
-
 # Object-level permissions.
 from guardian.shortcuts import assign_perm
-
 # For token creation.
 # Source: https://www.django-rest-framework.org/api-guide/authentication/#generating-tokens
 from rest_framework.authtoken.models import Token
 
+class Prefix(models.Model):
+    """BioCompute Prefix Model.
+
+     ...
+
+    Attributes
+    ----------
+    prefix: str
+        Prefix and primary key for this class
+    certifying_server: TextField, optional
+        Indicates server this prefix certified by.
+    certifying_key: TextField, optional
+        certifying key.
+    created: DateTimeField, optional
+        When was prefix created
+    created_by: User.username
+        User who created this prefix
+    description: str, optional
+        Descriptive string for this prefix
+    expires: DateTimeField, optional
+        Date and time indicating if the prefix expires
+    owner_group: Group.name,
+        Group that owns this prefix. Each prefix has exactly one group owner.
+    owner_user: User.username
+        User that owns this prefix. Each prefix has exactly one user owner.
+
+    Methods
+    -------
+    __str__(self)
+        String for representing the model (in Admin site etc.)
+    """
+
+    prefix = models.CharField(max_length=5, unique=True, primary_key=True)
+    certifying_server = models.TextField(blank = True, null = True)
+    certifying_key = models.TextField(blank = True,  null = True)
+    created = models.DateTimeField(default = timezone.now, blank = True, null = True)
+    created_by = models.ForeignKey(
+        User,
+        on_delete = models.CASCADE,
+        related_name = 'created_by',
+        to_field = 'username',
+        default='wheel')
+    description = models.TextField(blank = True, null = True)
+    expires = models.DateTimeField(blank = True, null = True)
+    n_objects = models.IntegerField(default=0)
+    owner_group = models.ForeignKey(Group, on_delete=models.CASCADE, to_field='name')
+    owner_user = models.ForeignKey(User, on_delete=models.CASCADE, to_field='username')
+
+    def __str__(self):
+        """
+        String for representing the BCO model (in Admin site etc.).
+        """
+        return str(self.prefix)
 
 # Generic BCO model
 class BCO(models.Model):
@@ -39,7 +88,7 @@ class BCO(models.Model):
     Attributes
     ----------
     object_id: str
-        BCO id and primary key for this DB
+        BCO id for this class
     contents: str, JSON
         Raw JSON contents of BCO
     object_class: str, optional
@@ -56,16 +105,21 @@ class BCO(models.Model):
         The state of the object, is it a draft, embargoed, or published
     last_update: DateTimeField
         When the draft was last updated
+
+    Methods
+    -------
+    __str__(self)
+        String for representing the model (in Admin site etc.)
     """
 
-    object_id = models.TextField(primary_key=True)
-    # Field is required.
+    id = models.AutoField(primary_key=True)
+    object_id = models.TextField()
     contents = models.JSONField()
     # TODO Embargo field.
     object_class = models.TextField(blank=True, null=True)
     owner_group = models.ForeignKey(Group, on_delete=models.SET_NULL, to_field='name', null=True)
     owner_user = models.ForeignKey(User, on_delete=models.SET_NULL, to_field='username', null=True)
-    prefix = models.ForeignKey(Prefix, on_delete=models.SET_NULL)
+    prefix = models.ForeignKey(Prefix, to_field='prefix', on_delete=models.PROTECT)
     schema = models.TextField()
     state = models.TextField()
     last_update = models.DateTimeField()
@@ -82,111 +136,35 @@ class GroupInfo(models.Model):
     anything relating to authentication.
     Delete group members on group deletion?
     """
-
+    group = models.OneToOneField(
+        Group,
+        on_delete=models.CASCADE,
+        to_field='name',
+        primary_key=True
+    )
     delete_members_on_group_deletion = models.BooleanField(default=False)
-
-    # The group
-    group = models.OneToOneField(Group, on_delete=models.CASCADE)
-    # Delete group members on group deletion?
-    delete_members_on_group_deletion = models.BooleanField(default=False)
-    # Description of the group
     description = models.TextField(blank = True)
-    # Expiration date
     expiration = models.DateTimeField(blank=True, null=True)
-
-    # The group
-    group = models.ForeignKey(Group, on_delete=models.CASCADE, to_field='name')
-
-    # Size limit for the number of members
     max_n_members = models.IntegerField(blank=True, null=True)
-
-    # Which user owns it?
     owner_user = models.ForeignKey(User, on_delete=models.CASCADE, to_field='username')
-
-
-# Generic meta data model
-# TODO: rename to prefix_meta
-class prefix_table(models.Model):
-    # The number of objects for a given prefix.
-
-    # Field is required.
-    n_objects = models.IntegerField()
-
-    # Which prefix the object falls under.
-
-    # Field is required.
-    prefix = models.CharField(primary_key=True, max_length=5)
-
-    def __str__(self):
-        """String for representing the BCO model (in Admin site etc.)."""
-        return self.prefix
 
 # For registering new users.
-class new_users(models.Model):
-    # Instead of using the User model, just use
-    # a crude table to store the temporary information
-    # when someone asks for a new account.
-    email = models.EmailField()
+class NewUsers(models.Model):
+    """
+    Instead of using the User model, just use
+    a crude table to store the temporary information
+    when someone asks for a new account.
+    In case we are writing back to UserDB.
+    Which host to send the activation back to (i.e. UserDB).
+    Issue with time zone, so implement the fix.
+    Source: https://stackoverflow.com/a/32411560
+    """
+
+    email = models.EmailField(primary_key=True)
     temp_identifier = models.TextField(max_length=100)
-    # In case we are writing back to UserDB.
     token = models.TextField(blank=True, null=True)
-
-    # Which host to send the activation back to (i.e. UserDB).
     hostname = models.TextField(blank=True, null=True)
-    # Issue with time zone, so implement the fix.
-    # Source: https://stackoverflow.com/a/32411560
     created = models.DateTimeField(default=timezone.now)
-
-
-# Link prefixes to groups and users.
-
-# Be careful about related_name.
-# Source: https://stackoverflow.com/questions/53651114/using-same-foreign-key-twice-in-a-model-in-django-as-different-fields
-class prefixes(models.Model):
-    
-    # Which server is this prefix certified with?
-    certifying_server = models.TextField(blank = True, null = True)
-
-    # What is the certifying key?
-    certifying_key = models.TextField(blank = True, null = True)
-    
-    # When was it created?
-    created = models.DateTimeField(
-        default = timezone.now,
-        blank = True,
-        null = True
-    )
-
-    # Who created it?
-    created_by = models.ForeignKey(
-        User,
-        on_delete = models.CASCADE,
-        related_name = 'created_by',
-        to_field = 'username',
-        default='wheel'
-    )
-
-    # Description.
-    description = models.TextField(blank = True, null = True)
-
-    # When does it expire?
-    expires = models.DateTimeField(
-        blank = True,
-        null = True
-    )
-
-    # Each prefix has exactly one group owner.
-    owner_group = models.ForeignKey(Group, on_delete=models.CASCADE, to_field='name')
-
-    # Each prefix has exactly one user owner.
-    owner_user = models.ForeignKey(User, on_delete=models.CASCADE, to_field='username')
-
-    # The actual prefix.
-    prefix = models.CharField(max_length=5)
-    
-    def __str__(self):
-        """String for representing the BCO model (in Admin site etc.)."""
-        return self.prefix
 
 # def get_first_name(self):
 #     return self.first_name
@@ -208,25 +186,23 @@ class prefixes(models.Model):
 
 
 # Link user creation to groups.
-@receiver(
-    post_save,
-    sender=User
-)
-def associate_user_group(
-    sender, 
-    instance, 
-    created, 
-    **kwargs
-):
+@receiver(post_save, sender=User)
+def associate_user_group(sender, instance, created, **kwargs):
+    """
+    Create a group for this user.
+    Source: https://stackoverflow.com/a/55206382/5029459
+    Automatically add the user to the BCO drafters and publishers groups,
+    if the user isn't anon or the already existent bco_drafter or bco_publisher.
+    """
+
     if created:
-        # Create a group for this user.
-        # Source: https://stackoverflow.com/a/55206382/5029459
-        Group.objects.create(name=instance)
+        try:
+            Group.objects.get(name=instance)
+        except Group.DoesNotExist:
+            Group.objects.create(name=instance)
         group = Group.objects.get(name=instance)
         group.user_set.add(instance)
 
-        # Automatically add the user to the BCO drafters and publishers groups,
-        # if the user isn't anon or the already existent bco_drafter or bco_publisher.
         if instance.username not in ['anon', 'bco_drafter', 'bco_publisher']:
             User.objects.get(username=instance).groups.add(Group.objects.get(name='bco_drafter'))
             User.objects.get(username=instance).groups.add(Group.objects.get(name='bco_publisher'))
@@ -289,29 +265,24 @@ def delete_group_perms(
 
 
 # Link prefix creation to permissions creation.
-@receiver(
-    post_save,
-    sender=prefixes
-)
-def create_permissions_for_prefix(
-        sender,
-        instance=None,
-        created=False,
-        **kwargs
-):
+@receiver(post_save, sender=Prefix)
+def create_permissions_for_prefix(sender, instance=None, created=False, **kwargs):
+    """
+    Test
+    """
+
     if created:
 
         # Check to see whether or not the permissions
         # have already been created for this prefix.
         try:
-
             # Create the macro-level, draft, and publish permissions.
             for perm in ['add', 'change', 'delete', 'view', 'draft', 'publish']:
                 Permission.objects.create(
                     name='Can ' + perm + ' BCOs with prefix ' + instance.prefix,
                     content_type=ContentType.objects.get(
                         app_label='api',
-                        model='BCO'
+                        model='bco'
                     ),
                     codename=perm + '_' + instance.prefix
                 )
@@ -345,72 +316,43 @@ def create_permissions_for_prefix(
             # The permissions already exist.
             pass
 
+@receiver(post_delete, sender=Prefix)
+def delete_permissions_for_prefix(sender, instance=None, **kwargs):
+    """
+    Link prefix deletion to permissions deletion.
+    No risk of raising an error when using
+    a filter.
+    """
 
-# Link prefix deletion to permissions deletion.
-@receiver(
-    post_delete,
-    sender=prefixes
-)
-def delete_permissions_for_prefix(
-        sender,
-        instance=None,
-        **kwargs
-):
-    # No risk of raising an error when using
-    # a filter.
-    Permission.objects.filter(
-        codename='add_' + instance.prefix
-    ).delete()
+    Permission.objects.filter(codename='add_' + instance.prefix).delete()
+    Permission.objects.filter(codename='change_' + instance.prefix).delete()
+    Permission.objects.filter(codename='delete_' + instance.prefix).delete()
+    Permission.objects.filter(codename='view_' + instance.prefix).delete()
+    Permission.objects.filter(codename='draft_' + instance.prefix).delete()
+    Permission.objects.filter(codename='publish_' + instance.prefix).delete()
 
-    Permission.objects.filter(
-        codename='change_' + instance.prefix
-    ).delete()
-
-    Permission.objects.filter(
-        codename='delete_' + instance.prefix
-    ).delete()
-
-    Permission.objects.filter(
-        codename='view_' + instance.prefix
-    ).delete()
-
-    Permission.objects.filter(
-        codename='draft_' + instance.prefix
-    ).delete()
-
-    Permission.objects.filter(
-        codename='publish_' + instance.prefix
-    ).delete()
-
-
-# --- Group info --- #
-
-
-# Link group creation to permission creation.
 @receiver(post_save, sender=GroupInfo)
-def create_group_perms(
-    sender, 
-    instance=None, 
-    created=False, 
-    **kwargs
-):
+def create_group_perms(sender, instance=None, created=False, **kwargs):
+    """
+    Link group creation to permission creation.
+    Check to see whether or not the permissions
+    have already been created for this prefix.
+
+    Create the permissions, then use group_info to give the group admin
+    the admin permissions.
+    Create the administrative permissions for the group.
+    Give the administrative permissions to the user
+    creating this group.
+    """
     if created:
-        # Check to see whether or not the permissions
-        # have already been created for this prefix.
         try:
-            # Create the permissions, then use group_info to give the group admin
-            # the admin permissions.
-            # Create the administrative permissions for the group.
-            for perm in ['add_members_' + Group.objects.get(id=instance.group_id).name,
-                         'delete_members_' + Group.objects.get(id=instance.group_id).name]:
+            for perm in ['add_members_' + Group.objects.get(name=instance.group_id).name,
+                         'delete_members_' + Group.objects.get(name=instance.group_id).name]:
                 Permission.objects.create(
                     name='Can ' + perm,
                     content_type=ContentType.objects.get(app_label='auth', model='group'),
                     codename=perm)
-                # Give the administrative permissions to the user
-                # creating this group.
-                import pdb; pdb.set_trace()
-                User.objects.get(id=instance.owner_user_id).user_permissions.add(Permission.objects.get(codename=perm))
+                User.objects.get(username=instance.owner_user_id).user_permissions.add(Permission.objects.get(codename=perm))
 
         except PermErrors.IntegrityError:
 
@@ -451,7 +393,7 @@ def create_object_perms(
                         name='Can ' + p,
                         content_type=ContentType.objects.get(
                             app_label='api',
-                            model='BCO'
+                            model='bco'
                         ),
                         codename=p
                     )
@@ -490,7 +432,7 @@ def create_object_perms(
                     name='Can publish a new version of an existing published object',
                     content_type=ContentType.objects.get(
                         app_label='api',
-                        model='BCO'
+                        model='bco'
                     ),
                     codename='publish_new_version_' + instance.object_id
                 )
